3693: Rename to ThemeContainer
