+++ native/src/App.tsx
@@ -24,7 +24,7 @@ import I18nProvider from './components/I18nProvider'
 import SnackbarContainer from './components/SnackbarContainer'
 import StaticServerProvider from './components/StaticServerProvider'
 import StatusBar from './components/StatusBar'
-import { ThemeContainer } from './components/ThemeContext'
+import ThemeContainer from './components/ThemeContainer'
 import TtsContainer from './components/TtsContainer'
 import { RoutesParamsType } from './constants/NavigationTypes'
 import { userAgent } from './constants/endpoint'
